+++ enclave-manager/web/src/components/enclaves/configuration/types.ts
@@ -1,0 +1,5 @@
+export type ConfigureEnclaveForm = {
+  enclaveName: string;
+  restartServices: boolean;
+  args: Record<string, any>;
+};