+++ scripts/docker-image-builder.sh
@@ -36,12 +36,22 @@ if [ -z "${image_tags}" ]; then
     show_helptext_and_exit
 fi
 
+uname_arch=$(uname -m)
+
+architecture="amd64"
+
+if [ "$uname_arch" == "x86_64" ] || [ "$uname_arch" == "amd64" ]; then
+    architecture="amd64"
+elif [ "$uname_arch" == "aarch64" ] || [ "$uname_arch" == "arm64" ]; then
+    architecture="arm64"
+fi
+
 # Argument processing
 if "${push_to_registry_container}"; then
   buildx_platform_arg='linux/arm64/v8,linux/amd64'
   push_flag='--push'
 else
-  buildx_platform_arg='linux/amd64' # TODO: infer the local arch if that's reasonable
+  buildx_platform_arg="linux/${architecture}"
   push_flag='--load'
 fi
 echo "Building docker image for architecture '${buildx_platform_arg}' with flag '${push_flag}'"